+++ src/panels/CreatePostmanCollectionHtmlObject.ts
@@ -1,5 +1,6 @@
 import * as vscode from "vscode";
-import {  dropdownOptions, toBoolean, uniqueArray, uniqueSort, arrayFrom1, arrayFrom0, nth} from "../utilities/functions";
+import {  dropdownOptions, toBoolean, uniqueArray, uniqueSort, arrayFrom1, arrayFrom0, nth, checkedString, valueString} from "../utilities/functions";
+import { ScenarioGridObject } from "../utilities/ScenarioGridObject";
 
 // fixed fields indices
 const carrierIndex = 0;
@@ -28,7 +29,7 @@ export class CreatePostmanCollectionHtmlObject {
     private _carrierCodes: {carrier: string,  carriercode: string }[] = [],
     private _scenarioFieldValues: string[],
     private _availableScenarios: string[], 
-    private _modularElements: string[],
+    private _modularElementsWithParents: {parent:string, element:string}[],
     private _independent: boolean,
     private _modularValue: boolean
   ) { }
@@ -40,6 +41,16 @@ export class CreatePostmanCollectionHtmlObject {
     const codiconsUri = this._uris[1];
     const mainUri     = this._uris[2];
     const styleUri    = this._uris[3];
+
+    // get scenario grid object
+    let scenarioGrid: ScenarioGridObject = new ScenarioGridObject(
+      this._availableScenarios, 
+      this._modularElementsWithParents, 
+      this._scenarioFieldValues, 
+      this._modularValue, 
+      +this._fieldValues[nofScenariosIndex], 
+      nofScenariosIndex
+    );
 
     // define panel HTML
     let html =  /*html*/`
@@ -64,38 +75,38 @@ export class CreatePostmanCollectionHtmlObject {
           </section>
 				</div>
 
-        <section class="${this._ifDependent('rowsingle')}${this._ifIndependent('row32')}">
+        <section class="${this._ifDependent('rowsingle')}${this._ifIndependent('rowflex')}">
 
           <section class="rowsingle">
             <section class="component-container">
+              <div class="component-sub-container">
 
+                <section class="component-example">
+                  <div class="floatleft">
+                    <vscode-checkbox id="independent" class="independent" ${checkedString(this._independent)}>Independent of existing integrations</vscode-checkbox>
+                  </div>
+                  <div class="floatleft">
+                    ${this._getRefreshButton()}
+                  </div>
+                </section>
 
-              <section class="component-example">
-                <div class="floatleft">
-                  <vscode-checkbox id="independent" class="independent" ${this._checkedString(this._independent)}>Independent of existing integrations</vscode-checkbox>
-                </div>
-                <div class="floatleft">
-                  ${this._getRefreshButton()}
-                </div>
-              </section>
+                <vscode-divider role="separator"></vscode-divider>
 
-              <vscode-divider role="separator"></vscode-divider>
+                <h2>Carrier</h2>
 
-              <h2>Carrier</h2>
+                ${this._ifIndependent(this._getIndependentCarrierFolderStructureGrid())}
+                ${this._ifDependent(this._getCarrierFolderStructureGrid())}
 
-              ${this._ifIndependent(this._getIndependentCarrierFolderStructureGrid())}
-              ${this._ifDependent(this._getCarrierFolderStructureGrid())}
+                ${this._getCarrierDetailsGrid()}
 
-              ${this._getCarrierDetailsGrid()}
+                <vscode-divider role="separator"></vscode-divider>
 
-              <vscode-divider role="separator"></vscode-divider>
-
-              ${this._getHeadersGrid()} 
-
+                ${this._getHeadersGrid()} 
+              </div>
             </section> 
           </section>
 
-            ${this._ifIndependent(this._getScenariosGrid(this._availableScenarios, this._modularElements))}
+            ${this._ifIndependent(scenarioGrid.getHtml())}
         </section>
 
 			</body>
@@ -105,48 +116,12 @@ export class CreatePostmanCollectionHtmlObject {
     return html;
   }
 
-  private _checkedString(checked: boolean): string {
-    let outString: string = '';
-    if (checked) {
-      outString = 'checked';
-    }
-
-    return outString;
-  }
-  
-  private _valueString(string: string): string {
-    let outString = '';
-    if (string !== undefined && string !== "") {
-      outString = `value="${string}"`;
-    }
-    return outString;
-  }
-
-  private _hiddenString(ifTrue: boolean): string {
-    let hiddenString = 'hidden';
-    if (ifTrue) {
-      hiddenString = '';
-    }
-
-    return hiddenString;
-  }
-
   private _ifIndependent(content: string): string {
-    let outString = '';
-    if (this._independent) {
-      outString = content;
-    }
-
-    return outString;
+    return this._independent ? content : '';
   }
 
   private _ifDependent(content: string): string {
-    let outString = '';
-    if (!this._independent) {
-      outString = content;
-    }
-
-    return outString;
+    return this._independent ? '' : content;
   }
 
   private _getCreateButton(): string {
@@ -174,15 +149,15 @@ export class CreatePostmanCollectionHtmlObject {
       <section class="component-example">
 
         <p>Folder structure:    <b>carrier / api-name / module</b></p>
-        <vscode-dropdown id="carriername" class="dropdown" index="${carrierIndex}" ${this._valueString(this._fieldValues[carrierIndex])} position="below">
+        <vscode-dropdown id="carriername" class="dropdown" index="${carrierIndex}" ${valueString(this._fieldValues[carrierIndex])} position="below">
           ${dropdownOptions(this._carriers)}
         </vscode-dropdown>
         /
-        <vscode-dropdown id="carrierapiname" class="dropdown" index="${apiIndex}" ${this._valueString(this._fieldValues[apiIndex])} position="below">
+        <vscode-dropdown id="carrierapiname" class="dropdown" index="${apiIndex}" ${valueString(this._fieldValues[apiIndex])} position="below">
           ${dropdownOptions(this._apis)}
         </vscode-dropdown>
         /
-        <vscode-dropdown id="modulename" class="dropdown" index="${moduleIndex}" ${this._valueString(this._fieldValues[moduleIndex])} position="below">
+        <vscode-dropdown id="modulename" class="dropdown" index="${moduleIndex}" ${valueString(this._fieldValues[moduleIndex])} position="below">
           ${dropdownOptions(this._modules)}
         </vscode-dropdown>
 
@@ -198,14 +173,14 @@ export class CreatePostmanCollectionHtmlObject {
 
         <div class="floatleft">
           <p>Carrier</p>
-          <vscode-dropdown id="carriername" class="dropdownfield" index="${carrierIndex}" ${this._valueString(this._fieldValues[carrierIndex])} position="below">
+          <vscode-dropdown id="carriername" class="dropdownfield" index="${carrierIndex}" ${valueString(this._fieldValues[carrierIndex])} position="below">
             ${dropdownOptions(this._carriers)}
           </vscode-dropdown>
         </div>
 
         <div class="floatleftnopadding">
           <p>Module</p>
-          <vscode-dropdown id="modulename" class="dropdownfield" index="${moduleIndex}" ${this._valueString(this._fieldValues[moduleIndex])} position="below">
+          <vscode-dropdown id="modulename" class="dropdownfield" index="${moduleIndex}" ${valueString(this._fieldValues[moduleIndex])} position="below">
             ${dropdownOptions(this._modules)}
           </vscode-dropdown>
         </div>
@@ -221,21 +196,21 @@ export class CreatePostmanCollectionHtmlObject {
       <section class="component-example">
         <div class="floatleft">
           <p>Carrier Code</p>
-          <vscode-text-field id="carriercode" class="field" index="${carrierCodeIndex}" ${this._valueString(this._fieldValues[carrierCodeIndex])} readonly></vscode-text-field>
+          <vscode-text-field id="carriercode" class="field" index="${carrierCodeIndex}" ${valueString(this._fieldValues[carrierCodeIndex])} readonly></vscode-text-field>
         </div>
         <div class="floatleft">
           <p>Account number</p>
-          <vscode-text-field id="accountnumber" class="field" index="${accountNumberIndex}" ${this._valueString(this._fieldValues[accountNumberIndex])} placeholder="123456" size="5"></vscode-text-field>
+          <vscode-text-field id="accountnumber" class="field" index="${accountNumberIndex}" ${valueString(this._fieldValues[accountNumberIndex])} placeholder="123456" size="5"></vscode-text-field>
         </div>
         <div class="floatleftnopadding">
           <p>CostCenter</p>
-          <vscode-text-field id="costcenter" class="field" index="${costCenterIndex}" ${this._valueString(this._fieldValues[costCenterIndex])} placeholder="000001" size="5"></vscode-text-field>
+          <vscode-text-field id="costcenter" class="field" index="${costCenterIndex}" ${valueString(this._fieldValues[costCenterIndex])} placeholder="000001" size="5"></vscode-text-field>
         </div>
       </section>
       
       <section class="component-example">
         <p>Select customer</p>
-        <vscode-dropdown id="company" class="dropdown" index="${companyIndex}" ${this._valueString(this._fieldValues[companyIndex])} position="below">
+        <vscode-dropdown id="company" class="dropdown" index="${companyIndex}" ${valueString(this._fieldValues[companyIndex])} position="below">
             ${dropdownOptions(this._companies)}
           </vscode-dropdown>
       </section>     
@@ -251,7 +226,7 @@ export class CreatePostmanCollectionHtmlObject {
 
           <section class="component-example">
             <p>Number of headers</p>
-            <vscode-dropdown id="nofheaders" class="dropdown" index="${nofHeadersIndex}" ${this._valueString(this._fieldValues[nofHeadersIndex])} position="below">
+            <vscode-dropdown id="nofheaders" class="dropdown" index="${nofHeadersIndex}" ${valueString(this._fieldValues[nofHeadersIndex])} position="below">
               ${dropdownOptions(arrayFrom1(10))}
             </vscode-dropdown>
           </section>
@@ -273,12 +248,12 @@ export class CreatePostmanCollectionHtmlObject {
 
       subHeaderName += /*html*/ `
         <section class="component-example">
-          <vscode-text-field id="headername${header}" index="${header}" ${this._valueString(hNames[header])} class="headername" placeholder="header name"></vscode-text-field>
+          <vscode-text-field id="headername${header}" index="${header}" ${valueString(hNames[header])} class="headername" placeholder="header name"></vscode-text-field>
         </section>`;
 
         subPHeaderValue += /*html*/ `
         <section class="component-example">
-          <vscode-text-field id="headervalue${header}" index="${header}" ${this._valueString(hValues[header] )} class="headervalue" placeholder="header value"></vscode-text-field>
+          <vscode-text-field id="headervalue${header}" index="${header}" ${valueString(hValues[header] )} class="headervalue" placeholder="header value"></vscode-text-field>
         </section>`;
     }
 
@@ -293,57 +268,6 @@ export class CreatePostmanCollectionHtmlObject {
           ${subPHeaderValue}
         </div>
       </section>`;
-
-    return html;
-  }
-
-  private _getScenariosGrid(scenarios: string[], modularElements: string[]): string {
-    let scenariosGrid = /*html*/ `    
-    <section class="rowsingle">
-      <section class="component-container">
-        <h2>Scenarios</h2>
-
-        <vscode-text-field id="modularelements" value="${modularElements.join(',')}" hidden></vscode-text-field>
-
-        <section class="component-example">
-          <vscode-checkbox id="modular" class="modular" ${this._checkedString(this._modularValue)}>Modular</vscode-checkbox>
-        </section>
-
-        <section class="component-example">
-          <p>Number of Scenarios</p>
-          <vscode-dropdown id="nofscenarios" class="dropdown" index="${nofScenariosIndex}" ${this._valueString(this._fieldValues[nofScenariosIndex] ?? 0)} position="below">
-            ${dropdownOptions(arrayFrom0(100))}
-          </vscode-dropdown>
-        </section>
-
-        ${this._scenarioInputs(+this._fieldValues[nofScenariosIndex], scenarios)}
-      </section>
-    </section>`;
-
-    return scenariosGrid;
-  }
-
-  private _scenarioInputs(nofScenarios: number, scenarios: string[]): string {
-    let html: string = ``;
-
-    for (let scenario = 0; scenario < +nofScenarios; scenario++) {
-
-      let scenarioInputField: string = '';
-      if (this._modularValue) {
-        scenarioInputField = /*html*/ `<vscode-text-field id="scenario${scenario}" index="${scenario}" ${this._valueString(this._scenarioFieldValues[scenario])} class="scenariofield" placeholder="${(scenario + 1) + nth(scenario + 1)} scenario name..."></vscode-text-field>`;
-      } else {
-        scenarioInputField = /*html*/ `
-          <vscode-dropdown id="scenario${scenario}" index="${scenario}" ${this._valueString(this._scenarioFieldValues[scenario])} class="scenariofield" position="below">
-            <vscode-option></vscode-option>  
-            ${dropdownOptions(scenarios)}
-          </vscode-dropdown>`;
-      }
-
-      html += /*html*/`
-        <section class="component-example">
-          ${scenarioInputField}
-        </section>`;
-    }
 
     return html;
   }