+++ src/panels/CreatePostmanCollectionPanel.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import { getUri, getWorkspaceFile, getWorkspaceFiles, startScript, cleanPath, parentPath, uniqueSort, toBoolean, isEmpty, getAvailableIntegrations, getFromScript, getAvailableScenarios, getModularElements} from "../utilities/functions";
+import { getUri, getWorkspaceFile, getWorkspaceFiles, startScript, cleanPath, parentPath, uniqueSort, toBoolean, isEmpty, getAvailableIntegrations, getFromScript, getAvailableScenarios, getModularElements, getModularElementsWithParents} from "../utilities/functions";
 import * as fs from 'fs';
 import { CreatePostmanCollectionHtmlObject } from "./CreatePostmanCollectionHtmlObject";
 import { create } from "domain";
@@ -28,7 +28,7 @@ export class CreatePostmanCollectionPanel {
   private _modules: string[] = [];
   private _scenarioFieldValues: string[] = [];
   private _availableScenarios: string[] = [];
-  private _modularElements: string[] = [];
+  private _modularElementsWithParents: {parent:string, element:string}[] = [];
   private _independent: boolean = false;
   private _modularValue: boolean = false;
 
@@ -55,6 +55,7 @@ export class CreatePostmanCollectionPanel {
     // predefine some fixed fields
     this._fieldValues[moduleIndex] = 'booking';
     this._fieldValues[nofHeadersIndex] = "3";
+    this._fieldValues[nofScenariosIndex] = "1";
     
     // pre-allocate headers array
     this._headers = new Array<{name: string, value: string}>(20);
@@ -86,9 +87,9 @@ export class CreatePostmanCollectionPanel {
   // initial rendering
   public static render(extensionUri: vscode.Uri, nofSteps: number, context: vscode.ExtensionContext) {
     if (CreatePostmanCollectionPanel.currentPanel) {
-      CreatePostmanCollectionPanel.currentPanel._panel.reveal(vscode.ViewColumn.Two);
+      CreatePostmanCollectionPanel.currentPanel._panel.reveal(vscode.ViewColumn.One);
     } else {
-      const panel = vscode.window.createWebviewPanel("create-postman-collection", "Create Postman collection", vscode.ViewColumn.Two, {
+      const panel = vscode.window.createWebviewPanel("create-postman-collection", "Create Postman collection", vscode.ViewColumn.One, {
         enableScripts: true
       });
 
@@ -192,8 +193,8 @@ export class CreatePostmanCollectionPanel {
                   case moduleIndex:
                     getAvailableScenarios(this._fieldValues[moduleIndex]).then(scen => {
                       this._availableScenarios = scen;
-                      getModularElements(this._fieldValues[moduleIndex]).then(elements => {
-                        this._modularElements = elements;
+                      getModularElementsWithParents(this._fieldValues[moduleIndex]).then(elementsWithParents => {
+                        this._modularElementsWithParents = elementsWithParents;
                         this._updateWebview(extensionUri);
                       });
                     });
@@ -484,16 +485,16 @@ export class CreatePostmanCollectionPanel {
     await this._getCompanies();
     await this._getRestUrls();
     await this._getCarrierCodes();
-    this._availableScenarios = await getAvailableScenarios(this._fieldValues[moduleIndex]);
-    this._modularElements    = await getModularElements(this._fieldValues[moduleIndex]);
+    this._availableScenarios          = await getAvailableScenarios(this._fieldValues[moduleIndex]);
+    this._modularElementsWithParents  = await getModularElementsWithParents(this._fieldValues[moduleIndex]);
   }
 
   private async _getWebviewContent(webview: vscode.Webview, extensionUri: vscode.Uri): Promise<string> {
     // define necessary extension Uris
     const toolkitUri = getUri(webview, extensionUri, ["node_modules", "@vscode", "webview-ui-toolkit", "dist", "toolkit.js"]);
     const codiconsUri = getUri(webview, extensionUri, ["node_modules", "@vscode", "codicons", "dist", "codicon.css"]);
-    const mainUri = getUri(webview, extensionUri, ["panels", "createpostmancollection", "main.js"]);
-    const styleUri = getUri(webview, extensionUri, ["panels", "createpostmancollection", "style.css"]);
+    const mainUri = getUri(webview, extensionUri, ["scripts", "createpostmancollection", "main.js"]);
+    const styleUri = getUri(webview, extensionUri, ["scripts", "createpostmancollection", "style.css"]);
 
     // initialize (first time)
     if (this._integrationObjects.length === 0) {
@@ -502,8 +503,8 @@ export class CreatePostmanCollectionPanel {
       await this._getRestUrls();
       await this._getCarrierCodes();
       this._initializeValues();
-      this._availableScenarios = await getAvailableScenarios(this._fieldValues[moduleIndex]);
-      this._modularElements    = await getModularElements(this._fieldValues[moduleIndex]);
+      this._availableScenarios          = await getAvailableScenarios(this._fieldValues[moduleIndex]);
+      this._modularElementsWithParents  = await getModularElementsWithParents(this._fieldValues[moduleIndex]);
     }
 
     // crop flexible header field values
@@ -527,7 +528,7 @@ export class CreatePostmanCollectionPanel {
       this._carrierCodes,
       this._scenarioFieldValues,
       this._availableScenarios,
-      this._modularElements,
+      this._modularElementsWithParents,
       this._independent,
       this._modularValue
     );